requirements.txt
Django==1.10.2
django-ckeditor==5.0.3
django-photologue==3.5.1
django-sortedm2m==1.3.2
ExifRead==2.1.2
facebook-sdk==2.0.0
oauthlib==1.0.3
Pillow==3.1.1
pytz==2016.6.1
requests==2.9.1
requests-oauthlib==0.6.1
twython==3.3.0
twython-django==1.5.2
